src/requirements.txt
asn1crypto==1.5.1
blinker==1.9.0
cbor2==5.6.5
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pycparser==2.22
pyOpenSSL==25.0.0
python-dotenv==1.0.1
typing_extensions==4.12.2
webauthn==2.4.0
Werkzeug==3.1.3
